Refactor campaign ID to campaign object in helpers
This commit simplifies the inputs for the 'get_campaign_details_for_user' function in the helpers file, changing campaign ID to an actual campaign object. This refactoring avoids repeatedly accessing the campaign ID attribute, thus making the function's code cleaner and more efficient. As a part of this change, the way the function is called in the campaign service is also updated.