chore: Update Canary version to 5.1.18
This commit includes changes for the Canary release - the version has been updated to 5.1.18 across various files, including api_version_canary.txt, build.json, and deployConfig.json. Docker images have been redirected to point to the 'api-canary' version in docker-compose.yml. This is set to update the canary environment with the new version.