Feat: Refactor server and celery running conditions, add Slack notifications
The conditions to determine if the server and celery should run locally have been simplified to a single check against the IS_LOCAL flag. Moreover, a new feature is introduced: when a task is run, based on its status, a message is sent to Slack. This will help keep track of task execution in production environment.