:rocket: Dev deployment
Docker image for app, celery_worker, and celery_beat has been updated from api-canary to api-dev in docker-compose, and the build version has been updated to 5.2.3 from 5.2.2 in several files. It also ensures that tags are correctly labeled for the dev environment in the Makefile and sets IS_CANARY to false in deployConfig.json.