:ambulance: Adjust event filtering query in subdomain section
The adjustment in the event filtering query under the subdomain section of the store event API was made. The modification was to retrieve the 'community' object before filtering the events to ensure it gets the correct community id. This ensures that events are correctly filtered according to their association with a specific subdomain.