feat: Add site data and menu loading functionality
This commit includes the creation of new functions that will support loading essential initial site data and site menu items. It also includes the addition of extra verification in the handling section to ensure community_id and subdomain are provided. Various utilities to fetch viable menu items and feature flags for a community were added.