Added functionality to load menu items
Various functions and routes have been introduced to facilitate the loading of menu items. This includes the helper functions in the api_utils file to filter and prepare the menu items, as well as appropriate routes and handlers for these items. These additions will allow easier interaction with the menus within a given community context.