Integrate Slack notifications for task statuses
A change was made so that task status messages will now be sent to slack instead of being logged. The Slack webhook URL has been imported and utilized with the send_slack_message utility function. As a result, whenever a task gets completed, the status, either Succeeded or Failed, will be directly sent to the given Slack channel.