Refactor: Add another check to know whether user is from custom site or ME site
This commit introduces a host parameter into the `has_no_custom_website` and `get_viable_menu_items` methods, enabling the application to distinguish between different hosts. The implementation in the caller methods, specifically in the misc.py files for the API handlers and the store, has been updated to pass this host argument. This change allows the application to differentiate menu options based on the host URL which leads to a more tailored user experience.