refactor: Add 'is_link_external' and 'key' fields to CustomMenuItem model
Two new fields 'is_link_external' and 'key' have been added to the CustomMenuItem database model. Also, a function to save method is implemented, where if the key field is blank, an autogenerated key using the name of menu item will be created. A new migration file was generated to reflect these changes in the database schema.