feat: Add functionality to fetch internal links
A new function has been introduced to fetch a list of internal links, which can be specified for either footer menus or navbar menus. It has been integrated into the api services and handlers, allowing the links to be requested and received through a new endpoint. Error handling is also included.