feat: Added method to retrieve community logo link from menu in models.py
A new method called 'get_logo_link_from_menu' has been added in the database models. This method retrieves the community logo link from the published menu. The result of this method is also included in the json response representation of a community object. This new feature enhances the data provided about a community.