test: Add test function for loading menu in frontend portal
A new test function has been added in `test_custom_menu.py` to check the loading of the menu for the frontend portal. This function validates the success of the response, checks if the response data is a list, verifies the expected length of the response list and asserts the names of the first two items in the list.