Test: Add multiple test cases for custom menu operations (#1066)
* feat: Add multiple test cases for custom menu operations

Add tests for reset custom menu with multiple scenarios such as invalid payload, invalid id, normal user, super admin, and community admin. Also, added tests for list of admins function, internal link for nav and footer menu, and additional edge cases.

* test: Add test function for loading menu in frontend portal

A new test function has been added in `test_custom_menu.py` to check the loading of the menu for the frontend portal. This function validates the success of the response, checks if the response data is a list, verifies the expected length of the response list and asserts the names of the first two items in the list.

* refactor: Removed unused 'host' variable from misc.py and misc handlers

This change removes the unused 'host' variable from the file src/api/store/misc.py as well as from the misc handlers in src/api/handlers/misc.py. This simplification clears the confusion and cleans up the code.