feat: Add Rewiring America page and integrate with API
A new page 'Rewiring America' has been added to the application. This page includes a state calculator script from 'rewiringamerica.org'. Additionally, routing for the page has been defined in the miscellaneous API handler with the access to the environment variable for 'REWIRING_AMERICA_API_KEY'. Also, the endpoint for 'rewiring_america' in urls.py has been updated. The new page is also added to the portal main menu.