Refactor: Add fields_to_translate Attribute to Models Using CustomMeta in Database App (#1093)
* refactor: Add fields_to_translate attribute to models in models.py

This commit includes changes made to the models in the models.py file. Specifically, the attribute 'fields_to_translate' has been added to multiple classes in order to provide better i18n support.

* feat: Added a new class 'CustomMeta' in base_model.py

A new class 'CustomMeta' has been added in utils/base_model.py. The class is responsible for extracting 'fields_to_translate' from 'Meta', deleting it, and then injecting it directly into the class. This code change does not affect any existing features or functionalities.

* refactor: Added CustomMeta as metaclass to all model definitions in database.models.py

Updated all class model definitions in models.py to include CustomMeta as a metaclass. This update is expected to enhance the subclasses with improved customization options and flexibility. This refactor does not affect the actual functionality of any models, it's aimed at code structure enhancement.

* refactor: Updated codebase to use TranslationMeta class instead of CustomMeta

Removed the usage of CustomMeta class from models in database, and replaced with use of TranslationMeta class. Added TranslationMeta classes to models where translated fields are needed. This change is done to streamline the way translated fields in each model are defined and handled.