Fix Recurring Events Handling  (#1127)
This pull request addresses issues with the handling of recurring events. Previously, recurring events were updated whenever a user reloaded the community portal site through an endpoint. That endpoint was removed when we decided to load data from the API on an as-needed basis; hence, the start and end dates of recurring events stopped updating.